+++ frontend/src/pages/onboarding/AvatarForm.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const AvatarForm = () => <div>AvatarForm</div>;
