+++ frontend/src/pages/onboarding/ClimbingCertForm.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const ClimbingCert = () => <div>ClimbingCert</div>;
