+++ frontend/src/pages/onboarding/ClimbingGradesForm.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const ClimbingGrades = () => <div>ClimbingGrades</div>;
