+++ frontend/src/pages/onboarding/DetailsForm.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const DetailsForm = () => <div>DetailsForm</div>;
