+++ frontend/src/pages/onboarding/FavoriteGymsForm.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const FavoriteGyms = () => <div>FavoriteGyms</div>;
