+++ frontend/src/routes/paths.ts
@@ -18,10 +18,8 @@ export const PATH_PAGE = {
   page404: '/404',
 };
 
-// TODO: rename path, use numbering for onboarding steps?
 export const PATH_ONBOARDING = {
-  newuser: path(ROOTS_ONBOARDING, '/newuser'),
-  username: path(ROOTS_ONBOARDING, '/username'),
+  root: ROOTS_ONBOARDING,
 };
 
 export const PATH_DASHBOARD = {
